remove s after some models

+++ backend/db/migrations/20241103194635-create-review.js
@@ -22,7 +22,7 @@ module.exports = {
         allowNull: false,
       },
       stars: {
-        type: Sequelize.INTEGER(5),
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       reviewImage: {
